refactor: extract core creation
